+++ Changelog.md
@@ -5,6 +5,7 @@
 * Improve `--help` text
 * Improve error message when no tracks were found
 * Allow generated website to be placed in a non-root direcotory
+* Don't re-generate photos that are already present in the output directory
 
 ## v0.1.0 - 2026-05-02